added a comment regarding access to the example
I've encountered that it's not clear which address should be used to request the feed for the example, which is located in feed.yml. feed/example.rss? site/example.rss? Turns out, just an "example.rss". It took me ten minutes. I added a comment for those, like me, who were at a loss.